requirements.txt

Flask
gunicorn
requests
google-auth
psycopg2-binary
SQLAlchemy
Flask-SQLAlchemy